+++ x/feemarket/keeper/params.go
@@ -12,22 +12,11 @@ import (
 func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get(types.ParamsKey)
-	if len(bz) == 0 {
-		k.ss.GetParamSetIfExists(ctx, &params)
-	} else {
-		k.cdc.MustUnmarshal(bz, &params)
+	if bz == nil {
+		return params
 	}
-
-	// zero the nil params for legacy blocks
-	if params.MinGasPrice.IsNil() {
-		params.MinGasPrice = math.LegacyZeroDec()
-	}
-
-	if params.MinGasMultiplier.IsNil() {
-		params.MinGasMultiplier = math.LegacyZeroDec()
-	}
-
-	return
+	k.cdc.MustUnmarshal(bz, &params)
+	return params
 }
 
 // SetParams sets the fee market params in a single key