+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Renamed x/evm to x/vm
 - Renamed protobuf files from evmos to cosmos org
 - [\#83](https://github.com/cosmos/evm/pull/83) Remove base fee v1 from x/feemarket
+- [\#93](https://github.com/cosmos/evm/pull/93) Remove legacy subspaces
 
 ### API-Breaking